+++ src/utils/constants.ts
@@ -5,7 +5,7 @@ export const USE_MOCK_DATA = process.env.REACT_APP_USE_MOCK === 'true';
 export const ROWS_PER_PAGE_OPTIONS = [5, 10, 25];
 export const MINIMUM_SEARCH_CHARACTERS = 3;
 export const DEBOUNCE_INTERVAL_INPUT = 500;
-export const NAVIGATE_TO_PUBLIC_PUBLICATION_DURATION = 3000;
+export const NAVIGATE_TO_PUBLIC_REGISTRATION_DURATION = 3000;
 
 export const ORCID_BASE_URL = process.env.REACT_APP_ORCID_BASE_URL;
 export const ORCID_USER_INFO_URL = `${ORCID_BASE_URL}/oauth/userinfo`;
